Created new concern to improve override
