--- config/initializers/participatory_spaces_patch.rb
@@ -1,33 +1,0 @@
-# frozen_string_literal: true
-
-Rails.application.config.to_prepare do
-  Decidim::Assembly.class_eval do
-    def superspace
-      record = Decidim::Superspaces::SuperspacesParticipatorySpace.find_by(
-        participatory_space_id: id,
-        participatory_space_type: self.class.name
-      )
-      record.present? ? record.superspace : nil
-    end
-  end
-
-  Decidim::ParticipatoryProcess.class_eval do
-    def superspace
-      record = Decidim::Superspaces::SuperspacesParticipatorySpace.find_by(
-        participatory_space_id: id,
-        participatory_space_type: self.class.name
-      )
-      record.present? ? record.superspace : nil
-    end
-  end
-
-  Decidim::Conference.class_eval do
-    def superspace
-      record = Decidim::Superspaces::SuperspacesParticipatorySpace.find_by(
-        participatory_space_id: id,
-        participatory_space_type: self.class.name
-      )
-      record.present? ? record.superspace : nil
-    end
-  end
-end
